requirements.txt
flask
django
requests
streamlit
google-auth
google-auth-oauthlib